requirements.txt
django==4.0.3
djangorestframework==3.13.1
python-dotenv==0.20.0
pymongo==4.1.1
gunicorn==20.1.0
pydantic==1.9.1
python-jose==3.3.0
django-cors-headers==3.13.0
jdatetime==4.1.0